app/requirements.txt
# RAG Backend — Requirements
# Install: pip install -r requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── Plugin Adapters ──────────────────────────────────────
litellm>=1.40.0        # AI Gateway (replaces direct Groq/OpenRouter calls)
mlflow>=2.14.0         # Observability & session telemetry
